Вынес хедер в отдельный компонент
