+++ project/src/components/app/app.tsx
@@ -1,6 +1,6 @@
 import { Switch, Route, Router as BrowserRouter } from 'react-router-dom';
 import { AppRoute } from '../../const';
-import PrivateRoute from '../private-route/private-rout';
+import PrivateRoute from '../private-route/private-route';
 import MainPage from '../main-page/main-page';
 import LoginScreen from '../login-screen/login-screen';
 import FavoritesScreen from '../favorites-screen/favorites-screen';
