+++ project/src/components/favorites-screen-empty/favorites-screen-empty.tsx
@@ -1,55 +1,11 @@
-import { connect, ConnectedProps } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { AppRoute } from '../../const';
-import { logoutAction } from '../../store/api-actions';
-import { ThunkAppDispatch } from '../../types/action';
-import { State } from '../../types/state';
+import MainHeader from '../main-header/main-header';
 
-const mapStateToProps = ({ currentLogin }: State) => ({
-  currentLogin,
-});
-
-const mapDispatchToProps = (dispatch: ThunkAppDispatch) => ({
-  onLogout() {
-    dispatch(logoutAction());
-  },
-});
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
-
-type PropsFromRedux = ConnectedProps<typeof connector>;
-
-function FavoritesScreenEmpty({ onLogout, currentLogin }: PropsFromRedux): JSX.Element {
-
+function FavoritesScreenEmpty(): JSX.Element {
   return (
     <div className="page page--favorites-empty">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={AppRoute.Main}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">{currentLogin}</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to={AppRoute.Login}>
-                    <span className="header__signout" onClick={onLogout}>Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <MainHeader />
 
       <main className="page__main page__main--favorites page__main--favorites-empty">
         <div className="page__favorites-container container">
@@ -71,5 +27,4 @@ function FavoritesScreenEmpty({ onLogout, currentLogin }: PropsFromRedux): JSX.E
   );
 }
 
-export { FavoritesScreenEmpty };
-export default connector(FavoritesScreenEmpty);
+export default FavoritesScreenEmpty;
