+++ project/src/components/favorites-screen/favorites-screen.tsx
@@ -3,26 +3,18 @@ import { Link } from 'react-router-dom';
 import { AppRoute } from '../../const';
 import FavoritesScreenEmpty from '../favorites-screen-empty/favorites-screen-empty';
 import { State } from '../../types/state';
-import { ThunkAppDispatch } from '../../types/action';
 import { connect, ConnectedProps } from 'react-redux';
-import { logoutAction } from '../../store/api-actions';
+import MainHeader from '../main-header/main-header';
 
-const mapStateToProps = ({ offers, currentLogin }: State) => ({
+const mapStateToProps = ({ offers }: State) => ({
   offers,
-  currentLogin,
 });
 
-const mapDispatchToProps = (dispatch: ThunkAppDispatch) => ({
-  onLogout() {
-    dispatch(logoutAction());
-  },
-});
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
+const connector = connect(mapStateToProps);
 
 type PropsFromRedux = ConnectedProps<typeof connector>;
 
-function FavoritesScreen({ offers, onLogout, currentLogin }: PropsFromRedux): JSX.Element {
+function FavoritesScreen({ offers }: PropsFromRedux): JSX.Element {
   const favoriteLocations: Set<string> = new Set();
   offers.forEach((offer) => {
     if (offer.isFavorite) {
@@ -36,33 +28,7 @@ function FavoritesScreen({ offers, onLogout, currentLogin }: PropsFromRedux): JS
 
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to={AppRoute.Main}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">{currentLogin}</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to={AppRoute.Login}>
-                    <span className="header__signout" onClick={onLogout}>Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <MainHeader />
 
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
