+++ project/src/components/main-header/main-header.tsx
@@ -1,0 +1,65 @@
+import { connect, ConnectedProps } from 'react-redux';
+import { Link } from 'react-router-dom';
+import { AppRoute, AuthorizationStatus } from '../../const';
+import { logoutAction } from '../../store/api-actions';
+import { ThunkAppDispatch } from '../../types/action';
+import { State } from '../../types/state';
+
+const mapStateToProps = ({ authorizationStatus, currentLogin }: State) => ({
+  authorizationStatus,
+  currentLogin,
+});
+
+const mapDispatchToProps = (dispatch: ThunkAppDispatch) => ({
+  onLogout() {
+    dispatch(logoutAction());
+  },
+});
+
+const connector = connect(mapStateToProps, mapDispatchToProps);
+
+type PropsFromRedux = ConnectedProps<typeof connector>;
+
+function MainHeader({ authorizationStatus, onLogout, currentLogin }: PropsFromRedux): JSX.Element {
+  return (
+    <header className="header">
+      <div className="container">
+        <div className="header__wrapper">
+          <div className="header__left">
+            <Link className="header__logo-link header__logo-link--active" to={AppRoute.Main}>
+              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+            </Link>
+          </div>
+          <nav className="header__nav">
+            <ul className="header__nav-list">
+              {authorizationStatus === AuthorizationStatus.Auth ?
+                <>
+                  <li className="header__nav-item user">
+                    <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
+                      <div className="header__avatar-wrapper user__avatar-wrapper">
+                      </div>
+                      <span className="header__user-name user__name">{currentLogin}</span>
+                    </Link>
+                  </li>
+                  <li className="header__nav-item">
+                    <Link className="header__nav-link" to={AppRoute.Login}>
+                      <span className="header__signout" onClick={onLogout}>Sign out</span>
+                    </Link>
+                  </li>
+                </>
+                :
+                <li className="header__nav-item user">
+                  <Link className="header__nav-link" to={AppRoute.Login}>
+                    <span className="header__signout">Sign in</span>
+                  </Link>
+                </li>}
+            </ul>
+          </nav>
+        </div>
+      </div>
+    </header>
+  );
+}
+
+export { MainHeader };
+export default connector(MainHeader);
