+++ project/src/components/main-page-404/main-page-404.tsx
@@ -1,66 +1,11 @@
-import { connect, ConnectedProps } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../const';
-import { logoutAction } from '../../store/api-actions';
-import { ThunkAppDispatch } from '../../types/action';
-import { State } from '../../types/state';
+import { AppRoute } from '../../const';
+import MainHeader from '../main-header/main-header';
 
-const mapStateToProps = ({ authorizationStatus, currentLogin }: State) => ({
-  authorizationStatus,
-  currentLogin,
-});
-
-const mapDispatchToProps = (dispatch: ThunkAppDispatch) => ({
-  onLogout() {
-    dispatch(logoutAction());
-  },
-});
-
-const connector = connect(mapStateToProps, mapDispatchToProps);
-
-type PropsFromRedux = ConnectedProps<typeof connector>;
-
-function MainPage404({ authorizationStatus, onLogout, currentLogin }: PropsFromRedux): JSX.Element {
-
+function MainPage404(): JSX.Element {
   return (
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link header__logo-link--active" to={AppRoute.Main}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                {authorizationStatus === AuthorizationStatus.Auth ?
-                  <>
-                    <li className="header__nav-item user">
-                      <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
-                        <div className="header__avatar-wrapper user__avatar-wrapper">
-                        </div>
-                        <span className="header__user-name user__name">{currentLogin}</span>
-                      </Link>
-                    </li>
-                    <li className="header__nav-item">
-                      <Link className="header__nav-link" to={AppRoute.Login}>
-                        <span className="header__signout" onClick={onLogout}>Sign out</span>
-                      </Link>
-                    </li>
-                  </>
-                  :
-                  <li className="header__nav-item user">
-                    <Link className="header__nav-link" to={AppRoute.Login}>
-                      <span className="header__signout">Sign in</span>
-                    </Link>
-                  </li>}
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <MainHeader />
       <main className="page__main page__main--index page__main--index-empty">
         <h1 className="visually-hidden">Cities</h1>
 
@@ -82,5 +27,4 @@ function MainPage404({ authorizationStatus, onLogout, currentLogin }: PropsFromR
   );
 }
 
-export { MainPage404 };
-export default connector(MainPage404);
+export default MainPage404;
